refactor: rename utils to constants
